+++ samples/Drawing.hs
@@ -32,7 +32,7 @@ setup window = do
     url <- loadFile "image/png" (dir </> "game" </> "wizard-blue" <.> "png")
     img <- UI.img # set UI.src url
     
-    let positions = [(x,y) | x <- [0,20..300], y <- [0,20..300]] :: [(Int,Int)]
+    let positions = [(x,y) | x <- [0,20..300], y <- [0,20..300]] :: [UI.Point]
     on UI.click canvas $ const $ forM_ positions $
         \xy -> UI.drawImage img xy canvas