+++ src/Graphics/UI/Threepenny/Core.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE DeriveDataTypeable, RecordWildCards #-}
 {-# LANGUAGE RecursiveDo #-}
+{-# LANGUAGE FlexibleContexts #-}
 module Graphics.UI.Threepenny.Core (
     -- * Synopsis
     -- | Core functionality of the Threepenny GUI library.
@@ -41,6 +42,7 @@ module Graphics.UI.Threepenny.Core (
     (#), (#.),
     Attr, WriteAttr, ReadAttr, ReadWriteAttr(..),
     set, sink, get, mkReadWriteAttr, mkWriteAttr, mkReadAttr,
+    bimapAttr, fromObjectProperty,
     
     -- * Widgets
     Widget(..), element, widget,
@@ -48,11 +50,13 @@ module Graphics.UI.Threepenny.Core (
     -- * JavaScript FFI
     -- | Direct interface to JavaScript in the browser window.
     debug,
-    ToJS, FFI, ffi, JSFunction, runFunction, callFunction,
-    callDeferredFunction, atomic,
+    ToJS, FFI,
+    JSFunction, ffi, runFunction, callFunction,
+    HsFunction, ffiExport,
+    atomic,
     
     -- * Internal and oddball functions
-    fromProp, toElement,
+    fromJQueryProp, toElement,
     audioPlay, audioStop,
     
     ) where
@@ -80,7 +84,7 @@ import qualified Reactive.Threepenny as Reactive
 import qualified Graphics.UI.Threepenny.Internal.Driver  as Core
 import           Graphics.UI.Threepenny.Internal.Driver
     ( getRequestLocation
-    , callDeferredFunction, atomic, )
+    , atomic, )
 import           Graphics.UI.Threepenny.Internal.FFI
 import           Graphics.UI.Threepenny.Internal.Types   as Core
     ( Window, Config, defaultConfig, Events, EventData
@@ -343,6 +347,14 @@ callFunction fun = do
     window <- askWindow
     liftIO $ Core.callFunction window fun
 
+-- | Export the given Haskell function so that it can be called
+-- from JavaScript code.
+--
+-- TODO: At the moment, the function is not garbage collected.
+ffiExport :: IO () -> UI (HsFunction (IO ()))
+ffiExport fun = do
+    window <- askWindow
+    liftIO $ Core.newHsFunction window fun
 
 {-----------------------------------------------------------------------------
     Oddball
@@ -358,13 +370,6 @@ audioPlay el = runFunction $ ffi "%1.play()" el
 -- | Invoke the JavaScript expression @audioElement.stop();@.
 audioStop :: Element -> UI ()
 audioStop el = runFunction $ ffi "prim_audio_stop(%1)" el
-
--- Turn a jQuery property @.prop()@ into an attribute.
-fromProp :: String -> (JSON.Value -> a) -> (a -> JSON.Value) -> Attr Element a
-fromProp name from to = mkReadWriteAttr get set
-    where
-    set v el = runFunction $ ffi "$(%1).prop(%2,%3)" el name (to v)
-    get   el = fmap from $ callFunction $ ffi "$(%1).prop(%2)" el name
 
 {-----------------------------------------------------------------------------
     Layout
@@ -495,6 +500,17 @@ data ReadWriteAttr x i o = ReadWriteAttr
     , set' :: i -> x -> UI ()
     }
 
+instance Functor (ReadWriteAttr x i) where
+    fmap f = bimapAttr id f
+
+-- | Map input and output type of an attribute.
+bimapAttr :: (i' -> i) -> (o -> o')
+          -> ReadWriteAttr x i o -> ReadWriteAttr x i' o'
+bimapAttr from to attr = attr
+    { get' = fmap to . get' attr
+    , set' = \i' -> set' attr (from i')
+    }
+
 -- | Set value of an attribute in the 'UI' monad.
 -- Best used in conjunction with '#'.
 set :: ReadWriteAttr x i o -> i -> UI x -> UI x
@@ -533,6 +549,19 @@ mkReadAttr get = mkReadWriteAttr get (\_ _ -> return ())
 mkWriteAttr :: (i -> x -> UI ()) -> WriteAttr x i
 mkWriteAttr set = mkReadWriteAttr (\_ -> return ()) set
 
+-- | Turn a jQuery property @.prop()@ into an attribute.
+fromJQueryProp :: String -> (JSON.Value -> a) -> (a -> JSON.Value) -> Attr Element a
+fromJQueryProp name from to = mkReadWriteAttr get set
+    where
+    set v el = runFunction $ ffi "$(%1).prop(%2,%3)" el name (to v)
+    get   el = fmap from $ callFunction $ ffi "$(%1).prop(%2)" el name
+
+-- | Turn a JavaScript object property @.prop = ...@ into an attribute.
+fromObjectProperty :: (ToJS a, FFI (JSFunction a)) => String -> Attr Element a
+fromObjectProperty name = mkReadWriteAttr get set
+    where
+    set v el = runFunction  $ ffi ("%1." ++ name ++ " = %2") el v    
+    get   el = callFunction $ ffi ("%1." ++ name) el
 
 {-----------------------------------------------------------------------------
     Widget class